+++ src/views/certifications/components/accordion-with-mdx-content/index.tsx
@@ -2,60 +2,13 @@
  * Copyright (c) HashiCorp, Inc.
  * SPDX-License-Identifier: MPL-2.0
  */
-import NextImage, { ImageProps as NextImageProps } from 'next/image'
 import Accordion from 'components/accordion'
 import DevDotContent from 'components/dev-dot-content'
 import Image from 'components/image'
 import { ImageProps } from 'components/image/types'
+import ThemedMdxImage from '../themed-image'
 import { AccordionWithMdxContentProps, AccordionMdxItem } from './types'
 import s from './accordion-with-mdx-content.module.css'
-
-interface ThemedMdxImageProps {
-	src: {
-		light: string
-		dark: string
-	}
-	alt: NextImageProps['alt']
-	width: NextImageProps['width']
-	height: NextImageProps['height']
-	title?: NextImageProps['title']
-}
-
-function ThemedMdxImage({
-	alt,
-	src,
-	title,
-	width,
-	height,
-}: ThemedMdxImageProps) {
-	if (!src.light || !src.dark) {
-		throw new Error(
-			`[ThemedImage]: Please provide both 'dark' and 'light' src values`
-		)
-	}
-	return (
-		<>
-			<span data-hide-on-theme="dark">
-				<NextImage
-					src={src.light}
-					alt={alt}
-					title={title}
-					width={width}
-					height={height}
-				/>
-			</span>
-			<span data-hide-on-theme="light">
-				<NextImage
-					src={src.dark}
-					alt={alt}
-					title={title}
-					width={width}
-					height={height}
-				/>
-			</span>
-		</>
-	)
-}
 
 function MdxImage({
 	alt,
